# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NIAT SPI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance/search:
    get:
      operationId: searchStudents
      tags: [attendance]
      summary: Search students
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Student search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentSearchResult"
  /attendance/students/{studentId}/overview:
    get:
      operationId: fetchStudentOverview
      tags: [attendance]
      summary: Get student attendance overview
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
        - name: t
          in: query
          required: false
          description: SPI share token for public access (used when no staff session)
          schema:
            type: string
      responses:
        "200":
          description: Student overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentOverview"
        "403":
          description: Forbidden
  /attendance/students/{studentId}/subjects:
    get:
      operationId: fetchStudentSubjects
      tags: [attendance]
      summary: Get student subject-wise attendance
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
        - name: t
          in: query
          required: false
          description: SPI share token for public access (used when no staff session)
          schema:
            type: string
      responses:
        "200":
          description: Subject attendance list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectAttendance"
  /attendance/students/{studentId}/recent:
    get:
      operationId: fetchStudentRecentSessions
      tags: [attendance]
      summary: Get student recent sessions
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
        - name: t
          in: query
          required: false
          description: SPI share token for public access (used when no staff session)
          schema:
            type: string
      responses:
        "200":
          description: Recent sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionRecord"
  /attendance/students/{studentId}/quizzes:
    get:
      operationId: fetchStudentQuizzes
      tags: [attendance]
      summary: Get student quiz data
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
        - name: t
          in: query
          required: false
          description: SPI share token for public access (used when no staff session)
          schema:
            type: string
      responses:
        "200":
          description: Quiz data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentQuizzes"
...
components:
  schemas:
...
    StudentSearchResult:
      type: object
      required: [studentId, studentName]
      properties:
        studentId:
          type: string
        studentName:
          type: string
        instituteName:
          type: ["string", "null"]
        sectionName:
          type: ["string", "null"]
        attendancePct:
          type: ["number", "null"]
        presentCount:
          type: integer
        totalCount:
          type: integer
    StudentOverview:
      type: object
      required: [studentId, studentName, totalSessions, presentCount, absentCount, attendancePct, inRecovery, coursesInRecovery]
      properties:
        studentId:
          type: string
        studentName:
          type: string
        instituteName:
          type: ["string", "null"]
        sectionName:
          type: ["string", "null"]
        totalSessions:
          type: integer
        presentCount:
          type: integer
        absentCount:
          type: integer
        attendancePct:
          type: number
        inRecovery:
          type: boolean
        coursesInRecovery:
          type: integer
    SubjectAttendance:
      type: object
      required: [subjectTitle, present, total, pct, meetsRequirement]
      properties:
        subjectTitle:
          type: string
        present:
          type: integer
        total:
          type: integer
        pct:
          type: number
        meetsRequirement:
          type: boolean
    SessionRecord:
      type: object
      required: [date, sessionTitle, subjectTitle, attendanceStatus]
      properties:
        date:
          type: string
        sessionTitle:
          type: string
        subjectTitle:
          type: string
        attendanceStatus:
          type: string
        markingMethod:
          type: ["string", "null"]
    QuizItem:
      type: object
      required: [subjectTitle, title, score, maxScore, percentage, status]
      properties:
        subjectTitle:
          type: string
        title:
          type: string
        score:
          type: number
        maxScore:
          type: number
        percentage:
          type: number
        status:
          type: string
        date:
          type: ["string", "null"]
    QuizSummary:
      type: object
      required: [attempted, total, avgPct]
      properties:
        attempted:
          type: integer
        total:
          type: integer
        avgPct:
          type: number
    StudentQuizzes:
      type: object
      required: [classroomQuizzes, moduleQuizzes, classroomSummary, moduleSummary]
      properties:
        classroomQuizzes:
          type: array
          items:
            $ref: "#/components/schemas/QuizItem"
        moduleQuizzes:
          type: array
          items:
            $ref: "#/components/schemas/QuizItem"
        classroomSummary:
          $ref: "#/components/schemas/QuizSummary"
        moduleSummary:
          $ref: "#/components/schemas/QuizSummary"
...